wrote users_followers_commits_are_displayed_spec
